move rename and close event listeners to dedicated file
